back/pyproject.toml
[tool.poetry]
name = "saleor"
description = "A modular, high performance, headless e-commerce platform built with Python, GraphQL, Django, and React."
authors = [ "Saleor Commerce <hello@saleor.io>" ]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://saleor.io/"
repository = "https://github.com/mirumee/saleor"
documentation = "https://docs.saleor.io/"
version = "1.1.1"

  [tool.poetry.dependencies]
  python = "~3.9"
  babel = ">=2.8,<2.10"
  boto3 = "^1.17"
  braintree = ">=4.2,<4.12"
  dj-database-url = "^0"
  dj-email-url = "^1"
  django = "^3.2.4"
  django-countries = "^7.2"
  django-filter = "^2.3"
  django-measurement = "^3.0"
  django-mptt = "^0"
  django-phonenumber-field = ">=4,<6"
  django-prices = "^2.1"
  django-prices-openexchangerates = "^1.0.1"
  django-prices-vatlayer = "^1.1.0"
  django-versatileimagefield = "^2.1"
  draftjs-sanitizer = "^1.0.0"
  faker = ">=4.1,<9.0"
  freezegun = "^1"
  google-cloud-storage = "^1.29.0"
  google-i18n-address = "^2.3.5"
  google-measurement-protocol = "^1.0"
  graphene-django = "~2.13.0"
  graphene-federation = "^0.1.0"
  html-to-draftjs = "^1.0.1"
  markdown = "^3.1.1"
  maxminddb = ">=1.5.4,<3.0.0"
  petl = "1.7.4"
  opentracing = "^2.3.0"
  phonenumberslite = "^8.12.25"
  prices = "^1.0"
  psycopg2-binary = "^2.8.3"
  razorpay = "^1.2"
  requests = "^2.22"
  sentry-sdk = "^1.0"
  stripe = "^2.58.0"
  text-unidecode = "^1.2"
  weasyprint = ">=48,<=52.5"
  oauthlib = "^3.1"
  jaeger-client = "^4.5.0"
  openpyxl = "^3.0.3"
  django-cache-url = "^3.1.2"
  pyjwt = ">=1.7.1,<3.0.0"
  python-json-logger = ">=0.1.11,<2.1.0"
  pytimeparse = "^1.1.8"
  django-redis = "^5.0.0"
  Adyen = "^4.0.0"
  google-cloud-pubsub = ">=1.7,<3.0"
  gunicorn = "^20.0.4"
  pybars3 = "^0.9.7"
  html2text = "^2020.1.16"
  authorizenet = "^1.1.3"
  sendgrid = "^6.7.1"
  micawber = "^0.5.2"
  django-celery-beat = "^2.2.1"

    [tool.poetry.dependencies.celery]
    version = ">=4.4.5,<6.0.0"
    extras = [ "redis" ]

    [tool.poetry.dependencies.django-storages]
    version = "^1.11"
    extras = [ "google" ]

    [tool.poetry.dependencies.python-magic-bin]
    version = "^0.4.14"
    platform = "win32"

    [tool.poetry.dependencies.uvicorn]
    extras = [ "standard" ]
    version = ">=0.12.2,<0.14.0"

  [tool.poetry.dev-dependencies]
  black = "21.7b0"
  codecov = "^2.1.11"
  coverage = "^5.5"
  django-debug-toolbar = "^3.2"
  django-debug-toolbar-request-history = "^0"
  django-graphiql-debug-toolbar = "^0.1.4"
  django-extensions = "^3.1.2"
  flake8 = "^3.9.2"
  isort = "^5.8.0"
  pre-commit = "^2.13"
  pycodestyle = "^2.5"
  pydocstyle = "^6.1"
  pylint = "^2.8.3"
  pylint-celery = "^0"
  pylint-django = "^2.4.4"
  pylint-plugin-utils = "^0"
  pytest = "^6.2.3"
  pytest-celery = "^0.0.0"
  pytest-cov = "^2.12.1"
  pytest-django = "4.2.0"
  pytest-django-queries = "~1.2"
  pytest-mock = "^3.6.1"
  pytest-vcr = "^1.0.2"
  pytest-xdist = "^2.2.1"
  tox = "^3.23.1"
  transifex-client = "^0"
  mypy = "0.910"
  pywatchman = "^1.4.1"
  types-requests = "^2.25.0"
  types-pytz = "^2021.1.0"
  types-pkg-resources = "^0.1.3"
  types-python-dateutil = "^0.1.4"
  django-stubs = "^1.8.0"

[tool.black]
target_version = [ "py35", "py36", "py37", "py38" ]
include = "\\.pyi?$"
exclude = """
/(\\.git/
  |\\.eggs
  |\\.hg
  |__pycache__
  |\\.cache
  |\\.ipynb_checkpoints
  |\\.mypy_cache
  |\\.pytest_cache
  |\\.tox
  |\\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
"""
